feat: Add TotalViewer component
This commit adds a new component called TotalViewer. The TotalViewer component is responsible for rendering a total amount based on the provided totalList prop. It calculates the sum of the numbers in the totalList array and displays it in a formatted way. This component will be used to display the total amount in the application.